fixing up the environment and testing server endpoints
