+++ full-stack-practice/client/.eslintrc.cjs
@@ -14,5 +14,6 @@ module.exports = {
       'warn',
       { allowConstantExport: true },
     ],
+    '@typescript-eslint/no-explicit-any': 'off',
   },
-}
+};
