+++ full-stack-practice/client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app-ts",
+  "name": "client",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -12,8 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.12.0",
-    "react-router-dom": "^6.29.0"
+    "react-icons": "^5.5.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
