--- full-stack-practice/client/src/lib/data.ts
@@ -1,71 +1,0 @@
-export type Product = {
-  productId: number;
-  name: string;
-  imageUrl: string;
-  price: number;
-  shortDescription: string;
-  longDescription: string;
-};
-
-export const products: Product[] = [
-  {
-    productId: 11,
-    name: 'Shake Weight',
-    price: 2999,
-    imageUrl: '/images/shake-weight.jpg',
-    shortDescription:
-      'Dynamic Inertia technology ignites muscles in arms, shoulders, and chest.',
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-  {
-    productId: 42,
-    name: 'ShamWow',
-    price: 2595,
-    imageUrl: '/images/shamwow.jpg',
-    shortDescription:
-      "It's like a chamois, towel, and sponge, all in one! Soaks up to 10x it's weight in any liquid!",
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-  {
-    productId: 2,
-    name: 'Snuggie',
-    price: 2900,
-    imageUrl: '/images/snuggie.jpg',
-    shortDescription:
-      'Super-Soft Fleece with pockets! One Size fits all Adults! Keeps you Warm & Your Hands-Free!',
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-  {
-    productId: 89,
-    name: 'Wax Vac',
-    price: 999,
-    imageUrl: '/images/wax-vac.jpg',
-    shortDescription:
-      'Gentle way to remove ear wax. Safe and hygienic. Reduces the risk of painful infections.',
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-  {
-    productId: 121,
-    name: 'Ostrich Pillow',
-    price: 9900,
-    imageUrl: '/images/ostrich-pillow.jpg',
-    shortDescription:
-      'Create your own snugly space in the world and feel-good anywhere with the ultimate cocoon pillow.',
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-  {
-    productId: 36,
-    name: 'Tater Mitts',
-    price: 830,
-    imageUrl: '/images/tater-mitts.jpg',
-    shortDescription:
-      '8 Seconds is all you need with Tater Mitts. Quickly and easily prepare all your favorite potato dishes with Tater Mitts.',
-    longDescription:
-      "Lorem ipsum dolor amet fashion axe pour-over jianbing, adaptogen waistcoat tacos master cleanse pitchfork next level. Thundercats pour-over chartreuse 90's. Master cleanse hot chicken ennui offal. Freegan slow-carb offal hell of. Umami polaroid wolf slow-carb next level. Gentrify cardigan seitan, kombucha tacos chambray roof party typewriter man braid. Tote bag lo-fi hell of chia fam hammock.\n\nAesthetic photo booth la croix, vaporware leggings biodiesel man braid tumeric skateboard tousled slow-carb four dollar toast synth pabst pickled. Typewriter church-key chia slow-carb vice gochujang actually. Shoreditch austin woke hot chicken, single-origin coffee ugh affogato four loko green juice. Migas iPhone four dollar toast mustache.",
-  },
-];
