+++ full-stack-practice/client/src/lib/index.ts
@@ -1,3 +1,1 @@
-export * from './data';
-export * from './read';
 export * from './to-dollars';
